third_party/blink/web_tests/external/wpt/permissions/permissions-query-feature-policy-attribute.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] Permissions.state is "prompt" with allow="geolocation" in same-origin iframes.
  assert_equals: navigator.permissions.query("geolocation") expected "prompt" but got "denied"
[FAIL] Permissions.state is "prompt" with allow="geolocation" in cross-origin iframes.
  assert_equals: navigator.permissions.query("geolocation") expected "prompt" but got "denied"
[FAIL] Permission.state is "prompt" in same-origin iframes.
  assert_equals: navigator.permissions.query("geolocation") expected "prompt" but got "denied"
Harness: the test ran to completion.
